replace StringSwitch with switch for integer type.
